refactor(backend): Move CLUE_STATUS definition to constants
