+++ backend/src/utils/constants.js
@@ -1,0 +1,4 @@
+export const CLUE_STATUS = {
+    SOLVED: "solved",
+    SKIPPED: "skipped",
+};
